requirements-unix.txt
# --- core numeric / data stack ---
numpy
pandas
scipy
scikit-learn

# --- visualization ---
matplotlib
seaborn

# --- experiment helpers ---
tqdm          # progress bars
ipython
jupyter
ipykernel

# --- configuration / logging ---
python-dotenv
loguru

# --- optional: faster event loop (not available on Windows) ---
uvloop


pyarrow==20.0.0
aiohttp==3.11.18
transformers==4.51.3
asyncio=3.4.3
aiogram
openpyxl==3.1.5
langchain-community==0.3.25
sentence-transformers==4.1.0
chromadb==1.0.12
langchain-huggingface==0.3.0
email-validator==2.2.0
aiosmtplib==3.0.1
aiosqlite==0.21.0
